use a monitor-like output format with multiple lines
This changes the output to a monitor-like interface with multiple lines.
The old output printed all information in one line, which works fine if
there's not a lot of information.
However, with the addition of the prescaler the interface now has to
show:
- the measured frequency
- whether hold mode is on or off
- clock generator output
- digital filter
- prescaler configuration

Split the output into multiple lines:
1. measured frequency and hold mode
2. (empty line)
3. clock generator output
4. digital filter
5. prescaler

To make the output even easier to read this uses a monitor-like
interface, which clears the screen when a new sample was captured.